+++ src/DetailsView/components/left-nav/details-view-left-nav.tsx
@@ -3,7 +3,7 @@
 import { mapValues } from 'lodash';
 import * as React from 'react';
 
-import { AssessmentsProvider } from '../../../assessments/types/iassessments-provider';
+import { AssessmentsProvider } from '../../../assessments/types/assessments-provider';
 import { NamedSFC } from '../../../common/react/named-sfc';
 import { AssessmentStoreData } from '../../../common/types/store-data/assessment-result-data';
 import { FeatureFlagStoreData } from '../../../common/types/store-data/feature-flag-store-data';
