+++ src/common/assessment/requirement.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 import { keys, sortBy } from 'lodash';
 
-import { AssessmentsProvider } from '../../assessments/types/iassessments-provider';
+import { AssessmentsProvider } from '../../assessments/types/assessments-provider';
 import { Requirement } from '../../assessments/types/requirement';
 import { ManualTestStatusData, TestStepData } from '../types/manual-test-status';
 import { VisualizationType } from '../types/visualization-type';
