+++ src/tests/unit/tests/DetailsView/reports/get-assessment-summary-model.test.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
+import { AssessmentsProvider } from '../../../../../assessments/types/assessments-provider';
 import { Assessment } from '../../../../../assessments/types/iassessment';
-import { AssessmentsProvider } from '../../../../../assessments/types/iassessments-provider';
 import { ManualTestStatus, TestStepData } from '../../../../../common/types/manual-test-status';
 import { AssessmentStoreData } from '../../../../../common/types/store-data/assessment-result-data';
 import { OverviewSummaryReportModel } from '../../../../../DetailsView/reports/assessment-report-model';
