+++ src/tests/unit/tests/injected/drawing-controller.test.ts
@@ -3,7 +3,7 @@
 import { IMock, It, Mock, MockBehavior, Times } from 'typemoq';
 
 import { AssessmentsProviderImpl } from '../../../../assessments/assessments-provider';
-import { AssessmentsProvider } from '../../../../assessments/types/iassessments-provider';
+import { AssessmentsProvider } from '../../../../assessments/types/assessments-provider';
 import {
     VisualizationConfiguration,
     VisualizationConfigurationFactory,
